requirements.txt
# Attendancify Backend Requirements
# Install with: pip install -r requirements.txt

# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
sqlite3  # Built into Python

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data validation
pydantic==2.5.0
email-validator==2.1.0

# AI & Face Recognition
deepface==0.0.79
opencv-python==4.8.1.78
google-generativeai==0.3.2
Pillow==10.1.0
numpy==1.24.3

# TensorFlow for DeepFace compatibility
tf-keras==2.15.0

# HTTP requests
requests==2.31.0
httpx==0.25.2

# CORS support
python-cors==1.7.0

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
